requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh upgrade
agno==1.4.6
azure-ai-inference
aiohttp
azure-core>=1.30.0
azure-identity>=1.15.0
baidusearch
pycountry
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
curl-cffi==0.10.0
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
duckduckgo-search==8.0.1
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jiter==0.9.0
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.5
openai==1.78.0
pandas==2.2.3
peewee==3.18.1
pgvector==0.4.1
platformdirs==4.3.8
primp==0.15.0
protobuf==5.29.4
psycopg==3.2.7
psycopg-binary==3.2.7
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yfinance==0.2.59
